[#38] Refactor: BouncingDots 컴포넌트화
