+++ Beat100/Beat100/Resources/Presentation/Measuring/MeasuringCompleteView.swift
@@ -18,19 +18,9 @@ struct MeasuringCompleteView: View {
                 Text("측정완료")
                     .foregroundStyle(Color.white)
                     .font(.system(size: 24, weight: .heavy))
-                TimelineView(.animation) { timeline in
-                    let time = timeline.date.timeIntervalSinceReferenceDate
-                    HStack {
-                        ForEach(0..<3) { index in
-                            Circle()
-                                .frame(width: 16, height: 16)
-                                .foregroundStyle(index == 0 ? .beatPink : index == 1 ? .beatGradPink : .beatBlue)
-                                .offset(y: bounceOffset(for: index, time: time))
-                                .padding(.trailing, index < 2 ? 15 : 0)
-                        }
-                    }
-                    .frame(height: 130)
-                }
+                
+                BouncingDotsView(dotSize: 16)
+                
                 Text("리포트를 생성중입니다")
                     .foregroundStyle(Color.white)
                     .font(.system(size: 24, weight: .heavy))
@@ -48,19 +38,9 @@ struct MeasuringCompleteView: View {
                 Text("측정완료")
                     .foregroundStyle(Color.white)
                     .font(.system(size: 12, weight: .heavy))
-                TimelineView(.animation) { timeline in
-                    let time = timeline.date.timeIntervalSinceReferenceDate
-                    HStack {
-                        ForEach(0..<3) { index in
-                            Circle()
-                                .frame(width: 10, height: 10)
-                                .foregroundStyle(index == 0 ? .beatPink : index == 1 ? .beatGradPink : .beatBlue)
-                                .offset(y: bounceOffset(for: index, time: time))
-                                .padding(.trailing, index < 2 ? 10 : 0)
-                        }
-                    }
-                    .frame(height: 78)
-                }
+                
+                BouncingDotsView(dotSize: 10)
+                
                 Text("리포트를 생성중입니다")
                     .foregroundStyle(Color.white)
                     .font(.system(size: 12, weight: .heavy))
@@ -78,11 +58,6 @@ struct MeasuringCompleteView: View {
             
 #endif
         }
-    }
-    
-    func bounceOffset(for index: Int, time: TimeInterval) -> CGFloat {
-        let delay = Double(index) * 0.2
-        return sin((time + delay) * 4) * 5
     }
 }
